requirements/base.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
drf-spectacular
psycopg2-binary
gunicorn
redis
celery
pytest
pytest-django
